Returns a dataframe of the filters stored in the local database.

Returns a dataframe of the filters stored in the local database.

listRunsFilters(src)

Arguments

src

dplyr sqlite src, as returned by dplyr::src_sqlite()

Valeur de retour

a dataframe