Create a dataframe of simultaneous detections at multiple sites

Creates a dataframe consisting of detections of tags that are detected at two or more receiver at the same time.

simSiteDet(data)

Arguments

data
a selected table from .motus data, eg. "alltags", or a data.frame of detection data including at a minimum variables for motusTagID, recvDeployName, ts

Examples

# You can use either a selected tbl from .motus eg. "alltags", or a
# data.frame, instructions to convert a .motus file to all formats are below.

# download and access data from project 176 in sql format
# usename and password are both "motus.sample"
if (FALSE) sql.motus <- tagme(176, new = TRUE, update = TRUE)

# OR use example sql file included in `motus`
sql.motus <- tagme(176, update = FALSE, 
                   dir = system.file("extdata", package = "motus"))

# convert sql file "sql.motus" to a tbl called "tbl.alltags"
library(dplyr)
tbl.alltags <- tbl(sql.motus, "alltags") 

# convert the tbl "tbl.alltags" to a data.frame called "df.alltags"
df.alltags <- tbl.alltags %>% 
  collect() %>% 
  as.data.frame()

# To get a data.frame of just simultaneous detections from a tbl file
# tbl.alltags
simSites <- simSiteDet(tbl.alltags)

# To get a data.frame of just simultaneous detections from a dataframe
# df.alltags
simSites <- simSiteDet(df.alltags)