report how much new data motus has for a tag detection database

"new" means data not already in your local database.

tellme(projRecv, new = FALSE, dir = getwd())

Arguments

projRecv

integer scalar project code from motus.org, or character scalar receiver serial number

new

logical scalar: is this a new database? Default: FALSE You have to specify new=TRUE if you want a new local copy of the database to be created. Otherwise, tellme() assumes the database already exists, and will stop with an error if it cannot find it in the current directory. This is mainly to prevent inadvertent downloads of large amounts of data that you already have!

dir

path to the folder where you are storing databases Default: the current directory; i.e. getwd()

Valeur de retour

a named list with these items:

numBatches: number of batches having data for your database

numRuns: number of runs of tags detections with new data

numHits: number of new detections

numGPS: number of new GPS fixes covering the new detections

numBytes: estimated size of download, in bytes. This is an estimate of the uncompressed size, but data are gz-compressed for transfer, so the number of bytes you have to download is typically going to be smaller than this number by a factor of 2 or more.

Note

if you specify new=TRUE and the database does not already exist, it will be created (but empty).

Voir également