Filter alltags by activity

The activity table is used to identify batches with too much noise. Depending on the value of return these are filtered out, returned, or identified in the alltags view with the column probability. No changes to the database are made.

filterByActivity(
  src,
  return = "good",
  view = "alltags",
  minLen = 3,
  maxLen = 5,
  maxRuns = 100,
  ratio = 0.85
)

Arguments

src
src_sqlite object representing the database

return
Character. One of "good" (return only 'good' runs), "bad" (return only 'bad' runs), "all" (return all runs, but with a new probability column which identifies 'bad' (0) and 'good' (1) runs.

view
Character. Which view to use, one of "alltags" (faster) or "alltagsGPS" (with GPS data).

minLen
Numeric. The minimum run length to allow (equal to or below this, all runs are 'bad')

maxLen
Numeric. The maximum run length to allow (equal to or above this, all runs are 'good')

maxRuns
Numeric. The cutoff of number of runs in a batch (see Details)

ratio
Numeric. The ratio cutoff of runs length 2 to number of runs in a batch (see Details)

Value

tbl_SQLiteConnection

Details

Runs are identified by the following:

All runs with a length >= maxLen are GOOD

All runs with a length <= minLen are BAD

Runs with a length between minLen and maxLen are BAD IF both of the following is true:

belong to a batch where the number of runs is >= maxRuns

the ratio of runs with a length of 2 to the number of runs total is >= ratio

Examples

filterByActivity,Filter alltags by activity,Filter alltags by activity — filterByActivity,filterByActivity.Rd,character(0),filterByActivity,character(0),list(title = "Description", contents = "
The activity table is used to identify batches with too much noise.\nDepending on the value of return these are filtered out, returned, or\nidentified in the alltags view with the column probability. No changes\nto the database are made.
"),list(description = "The activity table is used to identify batches with too much noise.\nDepending on the value of return these are filtered out, returned, or\nidentified in the alltags view with the column probability. No changes\nto the database are made."),filterByActivity(
  src,
  return = "good",
  view = "alltags",
  minLen = 3,
  maxLen = 5,
  maxRuns = 100,
  ratio = 0.85
),list(list(name = "src", description = "
src_sqlite object representing the database
"), list(name = "return", description = "
Character. One of \"good\" (return only 'good' runs), \"bad\"\n(return only 'bad' runs), \"all\" (return all runs, but with a new\nprobability column which identifies 'bad' (0) and 'good' (1) runs.
"), list(name = "view", description = "
Character. Which view to use, one of \"alltags\" (faster) or\n\"alltagsGPS\" (with GPS data).
"), list(name = "minLen", 
    description = "
Numeric. The minimum run length to allow (equal to or below\nthis, all runs are 'bad')
"), list(name = "maxLen", description = "
Numeric. The maximum run length to allow (equal to or above\nthis, all runs are 'good')
"), list(name = "maxRuns", description = "
Numeric. The cutoff of number of runs in a batch (see Details)
"), list(name = "ratio", description = "
Numeric. The ratio cutoff of runs length 2 to number of runs in\na batch (see Details)
")),TRUE,

#' # download and access data from project 176 in sql format
# usename and password are both "motus.sample"
if (FALSE) sql.motus <- tagme(176, new = TRUE, update = TRUE)

# OR use example sql file included in `motus`
sql.motus <- tagme(176, update = FALSE, 
                   dir = system.file("extdata", package = "motus"))

tbl_good <- filterByActivity(sql.motus)
tbl_bad <- filterByActivity(sql.motus, return = "bad")
tbl_all <- filterByActivity(sql.motus, return = "all")

,list(list(title = "Value", contents = "
tbl_SQLiteConnection
", slug = "value"), list(title = "Details", contents = "
Runs are identified by the following:

\n

All runs with a length >= maxLen are GOOD

\n

All runs with a length <= minLen are BAD

\n

Runs with a length between minLen and maxLen are BAD IF both of the\nfollowing is true:

\n

belong to a batch where the number of runs is >= maxRuns

\n

the ratio of runs with a length of 2 to the number of runs total\nis >= ratio

\n

\n
\n", 
    slug = "details")),2021,list(name = "motus", version = "4.0.5", authors = "John Brzustowski, Denis LePage, Birds Canada"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "motus"),list(),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Reference\n
\n
\n  Changelog\n
", right = ""),list(.present = TRUE),list(version = "1.6.1"),FALSE