update entire metadata for receivers and tags from Motus server. Contrary to tagme, this function retrieves the entire set of metadata for tags and receivers, and not only those pertinent to the detections of your local file.

update entire metadata for receivers and tags from Motus server. Contrary to tagme, this function retrieves the entire set of metadata for tags and receivers, and not only those pertinent to the detections of your local file.

metadata(src, projectIDs = NULL, replace = TRUE, delete = FALSE)

Arguments

src
src_sqlite object representing the database (either tag or receiver)

projectIDs
optional integer vector of Motus projects IDs for which metadata should be obtained; default: NULL, meaning obtain metadata for all tags and receivers that your permissions allow.

replace
logical scalar; if TRUE (default), existing data replace the existing metadata with the newly acquired ones.

delete
logical scalar; Default = FALSE. if TRUE, the entire metadata tables are cleared (for all projects) before re-importing the metadata.

See also

Examples

# download and access data from project 176 in sql format
# usename and password are both "motus.sample"
if (FALSE) sql.motus <- tagme(176, new = TRUE, update = TRUE)

# OR use example sql file included in `motus`
sql.motus <- tagme(176, update = FALSE, 
                   dir = system.file("extdata", package = "motus"))
                   
# Add extended metadata to your file
if (FALSE) metadata(sql.motus)
  
# Access different metadata tables
library(dplyr)
tbl(sql.motus, "species")

#> # Source:   table<species> [?? x 6]
#> # Database: sqlite 3.35.5
#> #   [/Users/runner/work/_temp/Library/motus/extdata/project-176.motus]
#>       id english            french                scientific         group  sort
#>    <int> <chr>              <chr>                 <chr>              <chr> <int>
#>  1  4180 Semipalmated Plov… Pluvier semipalmé     Charadrius semipa… BIRDS    NA
#>  2  4670 Red Knot           Bécasseau maubèche    Calidris canutus   BIRDS    NA
#>  3  4680 Sanderling         Bécasseau sanderling  Calidris alba      BIRDS    NA
#>  4  4690 Semipalmated Sand… Bécasseau semipalmé   Calidris pusilla   BIRDS    NA
#>  5  4760 White-rumped Sand… Bécasseau à croupion… Calidris fuscicol… BIRDS    NA
#>  6  4780 Pectoral Sandpiper Bécasseau à poitrine… Calidris melanotos BIRDS    NA
#>  7  4820 Dunlin             Bécasseau variable    Calidris alpina    BIRDS    NA
#>  8  4980 American Woodcock  Bécasse d'Amérique    Scolopax minor     BIRDS    NA
#>  9 15560 Gray-cheeked Thru… Grive à joues grises  Catharus minimus   BIRDS    NA
#> 10 19050 White-crowned Spa… Bruant à couronne bl… Zonotrichia leuco… BIRDS    NA

tbl(sql.motus, "projs")

#> # Source:   table<projs> [?? x 5]
#> # Database: sqlite 3.35.5
#> #   [/Users/runner/work/_temp/Library/motus/extdata/project-176.motus]
#>      id name                          label    tagsPermissions sensorsPermissio…
#>   <int> <chr>                         <chr>              <int>             <int>
#> 1    47 Red Knot staging and migrati… Niles                 NA                NA
#> 2    57 Selva Colombia                Selva                 NA                NA
#> 3    82 Maine - American Woodcock     RBrownA…              NA                NA
#> 4   146 Neonicotinoid insecticides a… NEONICS               NA                NA
#> 5   176 Sample Data                   SampleD…              NA                NA

tbl(sql.motus, "tagDeps")

#> # Source:   table<tagDeps> [?? x 23]
#> # Database: sqlite 3.35.5
#> #   [/Users/runner/work/_temp/Library/motus/extdata/project-176.motus]
#>    deployID tagID projectID status  tsStart  tsEnd deferSec speciesID bandNumber
#>       <int> <int>     <int> <chr>     <dbl>  <dbl>    <int>     <int> <chr>     
#>  1     1077 10811        47 NA       1.41e9 1.44e9        0      4670 NA        
#>  2     1798 16011       176 NA       1.44e9 1.45e9        0      4690 NA        
#>  3     1818 17357       176 NA       1.44e9 1.45e9        0      4760 NA        
#>  4     1823 16035       176 NA       1.44e9 1.45e9        0      4690 NA        
#>  5     1824 16036       176 NA       1.44e9 1.45e9        0      4690 NA        
#>  6     1825 16037       176 NA       1.44e9 1.45e9        0      4690 NA        
#>  7     1826 16038       176 NA       1.44e9 1.45e9        0      4690 NA        
#>  8     1827 16039       176 NA       1.44e9 1.45e9        0      4690 NA        
#>  9     1832 16044       176 NA       1.44e9 1.45e9        0      4760 NA        
#> 10     1839 16047       176 NA       1.44e9 1.46e9        0      4670 NA        
#> # … with more rows, and 14 more variables: markerNumber <chr>,
#> #   markerType <chr>, latitude <dbl>, longitude <dbl>, elevation <dbl>,
#> #   comments <chr>, id <int>, bi <dbl>, tsStartCode <chr>, tsEndCode <chr>,
#> #   fullID <chr>, age <chr>, sex <chr>, test <int>

# Etc.