Download motus tag detections to a database

Download motus tag detections to a database

Usage

tagme(
  projRecv,
  update = TRUE,
  new = FALSE,
  dir = getwd(),
  countOnly = FALSE,
  forceMeta = FALSE,
  rename = FALSE,
  skipActivity = FALSE,
  skipNodes = FALSE,
  skipDeprecated = FALSE
)

Arguments

projRecv

Numeric project code from motus.org, or character receiver serial number.

update

Logical. Should any new data be downloaded and merged? Defaults to TRUE unless this is a new database (in which case you must specify update = TRUE explicitly).

new

Logical. Is this a new database? Default: FALSE You have to specify new = TRUE if you want a new local copy of the database to be created. Otherwise, this function assumes the database already exists, and will stop with an error if it cannot find it in the current directory. This is mainly to prevent inadvertent downloads of large amounts of data that you already have!

dir

Character. Path to the folder where you are storing databases Defaults to current directory; i.e. getwd().

countOnly

Logical. If FALSE, the default, then do requested database updates. Otherwise, return a count of items that would need to be transferred in order to update the database.

forceMeta

Logical. If TRUE, re-get metadata for tags and receivers, even if we already have them.

rename

Logical. If current SQLite database is of an older version, automatically rename that database for backup purposes and download the newest version. If FALSE (default), user is prompted for action.

skipActivity

Logical. Skip checking for and downloading activity? See ?activity for more details

skipNodes

Logical. Skip checking for and downloading nodeData? See ?nodeData for more details

skipDeprecated

logical. Skip fetching list of deprecated batches stored in deprecated. See ?deprecateBatches() for more details.

Value

a SQLite Connection for the (possibly updated) database, or a list of counts if countOnly = TRUE

See also

Examples

if (FALSE) {

# Create and open a local tag database for motus project 14 in the
# current directory

t <- tagme(14, new = TRUE)

# Update and open the local tag database for motus project 14;
# it must already exist and be in the current directory

t <- tagme(14, update = TRUE)

# Update and open the local tag database for a receiver;
# it must already exist and be in the current directory

t <- tagme("SG-1234BBBK4567", update = TRUE)

# Open the local tag database for a receiver, without
# updating it

t <- tagme("SG-1234BBBK4567")

# Open the local tag database for a receiver, but
# tell 'tagme' that it is in a specific directory

t <- tagme("SG-1234BBBK4567", dir = "Projects/gulls")

# update all existing project and receiver databases in `dir`

tagme()
}